add data in event superclass

+++ eventdispatcher/DataProcessor.java
@@ -1,4 +1,4 @@
 package eventdispatcher;
 public interface DataProcessor {
-    public void process(short[] data);
+    public void process(double[] data);
 }
